📇 Update README to contain data from previous MSL
The README is currently rather barren, mostly containing links to public
open-source documentation.

Originally, MSL was a separate repository that was started, but left
in an incomplete state -- but with a more detailed `README.md`, which
this repository is slowly integrating and becoming (hence the 2020
copyright but 2021 git commits).

The details from the old README.md have been ported over here. Much of
these details may be out-of-date or incorrect until the actual code
comes into play -- but should still be reasonably close to reality.