✨ Add utility for preventing nullability
Its frequently necessary to have APIs that work on pointers, such as
`void*` -- but where `nullptr` is never a valid value. It is beneficial
to be able to encode this in the type system to avoid repeat checks
for nullability, and to have greater confidence in the validity of the
system.

This adds a utility to do just this, called `not_null`.