📖 Remove all `\tparam` tags from concept docs
`clang-cl`'s `-Wall` is being treated like `-Weverything` and triggering
`-Wdocumentation` errors -- this time for `\tparam` uses on `concept`
definitions. Although the docs _are_ correct, it is erroring and
claiming that the `\tparam` tags are not associated to a template.

It's easier to simply remove these since they don't really add any
value anyway.