✨ Add `address` semantic quantity type
In many cases when working with memory, it can be useful to represent
a raw address in a numeric form, for the sole purpose of printing or
comparison.

Sometimes this may be represented as a `void*`; though such pointers
can be dangerously misused if they are converted back to a `T*` type
that is incorrect/improper.

To compensate for that, a separate `address` type has been added that
represents this as a 1-way translation to a numeric/printable/comparable
type.