✨ Add semantic type for alignment
Alignment in C++ is conventionally represented as a simple `std::size_t`
value. C++17 added a small distinguishing type for `std::align_val_t`
which is actually an `enum class` type, but for all intents and
purposes, it is effectively the same concept.

The issue with these representations is that the value is never
validated for correctness. In fact, it's possible to construct a bad
value without ever realizing it or meaning to, despite the fact that
alignment always requires a power-of-two.

To address this, a semantic type is added that always verifies
correctness first.