🔧 Add self-containment test module
This project was previously committing individual source files that
included the header in order to verify that each header was properly
"self-contained". Although this process works, it leads to high
signal-to-noise ratio which makes it complicated to determine which
sources actually contain information of value, and which ones are
just empty for this purpose.

Ultimately, this process was not worth the long-term issues it appears
to cause, and so this has been changed to a CMake module which
automatically generates the paths instead.