✨ Make `quantity<std::byte>` a `digital_quantity`
There currently exist two semantic unit types for counting:
`quantity<T>` and `digital_quantity`. The former is for counting any
`T` type -- effectively as a tagged `size_t` object. The latter is for
counting byte-like units in a semantic way similar to
`std::chrono::duration`. This becomes confusing once the type being
counted is a `std::byte`, since `quantity<std::byte>` is not also itself
a `digital_quantity`.

This updates the class to now make a special case that quantities of
bytes are actually just a subclass of `digital_quantity`