🔨 Remove range-support in `memory_block`
The `memory_block` was previously authored using C++20's ranges
support. There is nothing wrong with this approach, however it
appears that LLVM-14 appears to have broken `libc++` for at
least macOS by not shipping all of the ranges library -- despite
previously being supported in LLVM-13.

This could possibly be handled by selectively-enabling the
support, but ultimately this is just not worth the hassle. For the
time being, range support has just been removed.