🔨 Remove null `memory_block` support
The `memory_block` previously supported being created as a null
block. This support is not unreasonable overall, but does introduce
some design concerns overall by now requiring blocks to be tested for
being null before being assumed to be writeable. This design design
issue is, frankly, problematic and arbitrary.

By removing nullable construction, blocks become sequences of
*something*. They may still be empty if a range is truly empty, but the
blocks can never point to null -- which at least ensures that values
that a block with a large enough size should always be writeable as long
as the original pointer is valid.