✨ Add `intrusive_pointer_stack`
When working with memory pools and sequences, it can be important to
be able to concisely store sequences of free pools of memory that are
not always contiguous. The easiest approach to handle this is to store
the next available pool address into the currently stored pool.

To implement and support this, a unique and simple type has been added:
the `intrusive_pointer_stack`. Behaviorally, this is just a typical
stack; you can push a (valid non-null) pointer down, you can peek the
value from the top of the stack, or you can pop and remove the value.

This will become a building block in later allocator strategies.