✏️ Move uninitialized_storage to memory subsystem
The `uninitialized_storage` utilities were previously under the broad
`utility` subsystem, which was not ideal as this is a melting pot where
dependency ordering comes to die. This posed problems for subsystem
dependencies since some of the utilities wanted to depend on pointer
utilities, which in turn depends back on utilities.

There will be several abstractions that suffer this problem and have no
real location where it truly "belongs", but all deal with memory.
To solve this problem for now, any utilities that address raw memory
in some manner like this have been moved to a new `memory` subsytem.