✨ Add virtual memory support
When working with allocators, it's important to be able to reserve
large chunks of memory that can be split up into smaller more manageable
portions. Typically, this is done using virtual memory -- which is not
exposed by the standard library, and requires system APIs to do.

This adds support for allocating from the virtual memory to compose
larger allocator mechanisms.