✨ Add new `memory_block` utilities
`memory_block`s may frequently be used in contiguous segments of several
memory blocks -- for example, in the case of virtual memory.
This adds utilities to now allow both traversal of `memory_block`s, but
also total ordering of memory blocks.

These functionalities will now allow for blocks to be traversed, stored,
and detected in a sane way.