✨ Add `dynamic_storage` utility
For this library, it's desirable to specify certain demensions both
statically and, on occasion, at runtime as well. There is no ideal
mechanism in C++ that allows for achieving this -- at least, in the
standard library.

This introduces a lightweight mechanism that serves to _attempt_ to
solve this problem: the `dynamic_storage` (tentatively titled; may
be renamed to avoid confusion with the C++ terminology of 'dynamic
storage duration').

The `dynamic_storage` type behaves like a possibly-runtime value
where if `N` is a sentinel value, it can take the form of _any_
runtime value -- whereas if `N` is not the sentinel, it must be
statically specified. In this respect, it's nearly identical to the
manner in which  `std::span` operates in the presence of
`dynamic_extent`.

Mechanically, this now allows for classes to conditionally store a
size that can be specified at compile-time. This now allows for
compile-time or runtime pool mechanisms, for example, or fixed-sized
allocation schemes.