✨ Add `traversal_utilities`
In C++, it's not legal to treat a byte sequence capable of holding
`N` `T*` objects as a `T*` sequence, and to index into the sequence.

The initial pointer is "safely derived" and can be used only to either
constructs a real `T` object, or be converted back to the byte sequence.
However array arithmetic (e.g. `p + n` where `p` is a `T*`) is invalid
unless that pointer points to a true array of `T` objects (e.g. `T[N]`).

To worka round that, this introduces a collection of traversal
utilities. This makes the underlying assumption that allocators deal
with bytes, which is largely true (effectively 100% of the time).
Since a `T*` actually aliases bytes, it can legally be converted back
to bytes, offset by `N * sizeof(T)`, and converted back without violating
reachability or strict-aliasing. This makes the N'th pointer
'safely-derived' -- which is required for technical correctness.

These utilities abstract this concept, and also make the pedanticism
here optional -- which can be disabled with a preprocessor token.