🔨 Rework lifetime utility class
The `uninitialized_storage` class was never aptly named or placed
into a proper subsystem. It started in the `utilities` subsystem and
slowly navigated into the `memory` subsystem, neither of which
really accurately conveyed its belonging.

Since this is effectively a lifetime "pointer" utility class, this has
been renamed to `lifetime_utilities` and relocated into the `pointers`
subsystem, which is more applicable. Additionally, the classes have
also now been modified to only operate on not-null pointers instead.