✨ Add utilities for unaligned access
It's not uncommon when working with allocators to want to store
additional data before, or after, an allocation. In some cases, the
data might have a natural alignment that otherwise does not work well
with the alignment of the free-space left before or after the data.
In those cases, `std::memcpy` is conventionally necessary to be used.

`std::memcpy` is not always the easiest to work with in terms of its
readability (namely, the order of in/out appears ambiguous from the
call site). Additionally, the inputs can never be null, but nothing
actually tests or ensures this.

For this reason, a higher-order wrapper has been provided: the
`unaligned_utilities`. This is nothing more than a simple wrapper
around `mempcy` to give it a more coherent, and null-checked API.