🩹 Fix `alignment_of` implementation
The `alignment_of` utility in `pointer_utilities` was previously
violating teh `alignment` contract unintentionally by computing
the wrong alignment value.

The computation counts the right-trailing zeroes to determine the
alignment. This computation is actually incorrect; the proper alignment
value should be a digit shifted over by the number of zeroes (e.g.
with 0 zeroes, it's a 1-byte alignment; with 3 zeroes, it's an 8-byte
alignment, etc).

This fixes the computation so that the contract is no longer violated.