✨ Add `cell` abstraction
C++ pointers are fundamentally lacking when it comes to knowing its
own reachability or alignment in the type. This can easily lead to
issues where instructions that assume alignment are incorrectly used
and thus penalized or trapped, and code that assumes reachability is
violated and produces undefined behavior.

The proposed solution for this used in this library is to wrap this
into a smarter pointer-like abstraction knowing as `cell` objects.

Cells are effectively "fat pointers" when a size is not known until
runtime, but otherwise can conditionally code the entirety of the
reachability and alignment into its type.