🔨 Add new clang build configurations
Clang supports a collection of useful settings for sanitizer builds
and code-coverage builds which are not necessarily supported by
every compiler.

Although these could be managed by explicitly passing the flags to
the cmake configuration command, or done through a specific script
interaction, it's more convenient to expose these as dedicated
"configurations" in the build so that we can build these tests
independently from the normal builds.

This does just that by providing a new module to provide this
information.