🔨 Add cmake module to patch compile commands
Certain IDEs like VSCode make expectations about where the
`compile_commands.json` file exists. In particular, it expects this
file to be present in the `build/` directory, and not a sub-directory
of that build.

Although it's possible to reconfigure this path to expect a specific
subdirectory, this can be quite a hassle when the builtin CMake
plugin is already capable of configuring different paths easily. To
simplify this process, a new cmake `option` has been added that
will now determine if a "preset" is being used in CMake, which assumes
build directories to be under `build/`. If this is the case, it will
automatically patch the compile commands file without any hassle.